Rework private_vxc resource
Change sub-resources to TypeList. TypeSet is not suitable for this
application as the checks are based on the hash and from the looks of it
there is no way to have some attributes of the subresource cause updates
and some to force a new resource.

tl;dr - if the Set element's key (hash):
- does not change, then terraform considers no changes are to be made to
the the element
- does change, then terraform will replace the element (which will have
a new key) and if one if its attributes is set to `ForceNew`, the entire
resource is going to be recreated: even if the value of the `ForceNew`
attribute has not changed, its path in the state will have changed.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>