Refactor VxcService
Most of the different products on Megaport are VXCs. Their lifecycle is
controlled through the same endpoints on the Megaport API and the
process is more or less uniform, apart from the different payloads for
each type of VXC as presented on the web ui (eg.: private, partner,
cloud). Therefore, we can reuse the underlying CRUD methods in the
package and export methods specific to each type of VXC.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>