megaport_aws_vxc: do not update product_uid
According to the docs, when creating a VXC to a partner port the port
with `vxcPermitted == true` should be selected. This is the mechanism
that Megaport uses to load balance client connections.

However, the behaviour has recently changed in such a way that the API
will automatically load balance internally: requesting a connection to
some port might create a connection to a different one, which is
refelected on the data returned when reading the VXC data. The selected
port will be such that the resulting VXC is identical to the one
requested, in terms of connectivity. Moreover, it seems that the
`vxcPermitted` attribute of partner ports is not being updated.

The above is true for the AWS connections at least, so the workaround
is to avoid updating the `product_uid` attribute in the state (and
maintain the original configuration value). Additionally, we export
`connected_product_uid` which corresponds to the actual port where the
connection terminates. This has the effect of not producing a diff due
to the transparent load balancing mechanism but it will trigger an
update if the configuration changes.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>