megaport_aws_vxc: wait condition for deletions
After deleting a VXC, the parent port does not immediately make the VLAN id used
available. If we are to reuse the same VLAN id (eg.: when re-creating a VXC), we
should wait until it becomes available again or otherwise, Megaport will
automatically and silently chose a different id, which will generate a diff at
the next plan.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>