megaport_aws_vxc, megaport_gcp_vxc: fix imports
When importing a resource, we have no access to the config and cannot
properly setup the state with regard to the `product_uid` and
`connected_product_uid` attributes. This change leaves the `product_uid`
attribute empty if the value is unknown (which will happen when
importing any of these two resources). Additionally, it introduces a
DiffSuppressFunc which will force terraform to ignore changes to the attribute
if its value in the state is empty.

The result is that we have a clean plan after import, working around the
load balancing behaviour for cloud partner ports. However, it also means
that to modify the `product_uid` value of the B-End of a cloud VXC, the
resource will have to be tainted.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>