Always read MCR term and update the state
Even though MCRv2 does not have a "term" attribute, the API will always
return the value `1` for the respective field, which is also the default
value. We should not skip updating this field in the state for MCR v2,
when reading from the Megaport API.

If an MCRv2 is imported, the value will be `null` in state since there
is no configuration being parsed. Therefore, this generates a perpetual
diff. Updating based on the API will set the default value and prevent
this situation.

Signed-off-by: Dimitrios Karagiannis <dhkarag@gmail.com>